split up schemas
